// server/src/app/enums/typeCardRarity.ts
/**
 * Enum que indica la rareza de la carta
 * 
 * @COMMON indica que la rareza de la carta es comun
 * @UNCOMMON indica que la rareza de la carta es infrecuente
 * @RARE indica que la rareza de la carta es rara
 * @RARE_DOUBLE indica que la rareza de la carta es doble rara
 * @RARE_ILUSTRATION indica que la rareza de la carta es ilustración rara
 * @RARE_SPECIAL_ILUSTRATION indica que la rareza de la carta es ilustración especial rara
 * @RARE_ULTRA indica que la rareza de la carta es ultra rara
 * @RARE_HIPER indica que la rareza de la carta es hiper rara
 * @REVERSE_HOLO indica que la rareza de la carta es reverse holo
 * @PROMO indica que la rareza de la carta es promocional
 */
export enum TypeCardRarity {
  COMMON = 'Common',
  UNCOMMON = 'Uncommon',
  RARE = 'Rare',
  RARE_DOUBLE = 'Rare Double',
  RARE_ILUSTRATION = 'Rare Illustration',
  RARE_SPECIAL_ILUSTRATION = 'Rare Special Illustration',
  RARE_ULTRA = 'Rare Ultra',
  RARE_HIPER = 'Rare Hiper',
  REVERSE_HOLO = 'Reverse Holo',
  PROMO = 'Promo',
}